fast: Split DenseSparseDBM into two related interfaces.
Before the change, the DenseSparseDBM typeclass represented DBMs that

* are sparsity-aware
* can perform the usual DBM operations, like strong closure, in a
  sparsity-independent way
* can be decomposed into independent submatrices that are themselves
  DBMs with the above properties
* reintegrate their modified submatrices

With only one typeclass that can be instantiated by only a single DBM type,
we had that both the original DBM and its submatrices were forced to have all
of the above properties, because they were of the same type.

While being sparsity-aware and capable of performing DBM operations is necessary
for both the original DBM and its submatrices, the last two properties need
not (and actually must not) be supported by submatrices (it makes no sense and
only complicated implementations of the typeclass).

Therefore, we now have a base typeclass DenseSparse for DBM with the first two
properties, a SubMatrix typeclass for matrices that can be extracted from
another by specifying the subset of variable indices that should be covered,
and a Decomposable typeclass that is a binary relation between a type of
decomposable DBMs having the last two properties and a type of DBMs that can act
as its submatrices.